message_fetch: Move limit_query_to_range to zerver.lib.narrow.
